refactor: extract VM tests to dedicated files
Move 5 test modules (~2,000 lines) from vm/mod.rs to individual files:
parity_tests.rs, async_tests.rs, jit_tests.rs, schedule_watch_tests.rs,
must_ask_freeze_tests.rs. Reduces mod.rs from 2,058 to 50 lines.
